exec: add the hashed aggregator
hashedAggregator is an operator that performs an aggregation based on specified
grouping columns. This operator performs a blocking hash table build at the
beginning and then the grouped buckets are then passed as input to an
orderedAggregator. The output row ordering of this aggregator is arbitrary.

Release note: None

+++ pkg/sql/exec/coalescer_test.go
@@ -84,14 +84,14 @@ func BenchmarkCoalescer(b *testing.B) {
 
 	for colIdx := 0; colIdx < nCols; colIdx++ {
 		col := batch.ColVec(colIdx).Int64()
-		for i := 0; i < ColBatchSize; i++ {
+		for i := uint16(0); i < ColBatchSize; i++ {
 			col[i] = int64(i)
 		}
 	}
 
 	for _, nBatches := range []int{1 << 1, 1 << 2, 1 << 4, 1 << 8, 1 << 12, 1 << 16} {
-		b.Run(fmt.Sprintf("rows=%d", nBatches*ColBatchSize), func(b *testing.B) {
-			b.SetBytes(int64(8 * nBatches * ColBatchSize * nCols))
+		b.Run(fmt.Sprintf("rows=%d", nBatches*int(ColBatchSize)), func(b *testing.B) {
+			b.SetBytes(int64(8 * nBatches * int(ColBatchSize) * nCols))
 			b.ResetTimer()
 			for i := 0; i < b.N; i++ {
 				source := newRandomLengthBatchSource(batch)
